# src/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"TimeFrame": {
				"type": "string",
				"enum": [
					"MM",
					"YY",
					"30D",
					"60D",
					"90D",
					"AT"
				],
				"description": "Stringified time frame type"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "Wealthcare API",
		"version": "1.0.0",
		"description": "Restful API for Wealthcare app",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/account/{accountId}": {
			"get": {
				"operationId": "GetAccountById",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Get account by id",
				"tags": [
					"Account Controllers"
				],
				"security": [],
				"parameters": [
					{
						"description": "_ string",
						"in": "path",
						"name": "accountId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/account": {
			"post": {
				"operationId": "CreateNewAccount",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Account Controllers"
				],
				"security": [],
				"parameters": []
			}
		},
		"/dashboard": {
			"get": {
				"operationId": "GetDasboardData",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Default": {
										"value": {
											"netWorth": 35000000,
											"expense": 17500000,
											"income": 22600000,
											"chartData": {
												"netWorthGrowth": [
													4560000,
													12000000,
													18239000,
													22350000,
													25424000,
													4560000,
													12000000,
													18239000,
													22350000,
													25424000,
													35000000
												],
												"expenses": [
													{
														"id": "52907745-7672-470e-a803-a2f8feb52944",
														"name": "Food",
														"amount": 1560000
													},
													{
														"id": "52907745-7672-470e-a803-a2f8feb52944",
														"name": "Invest",
														"amount": 4000000
													},
													{
														"id": "52907745-7672-470e-a803-a2f8feb52944",
														"name": "Home",
														"amount": 6000000
													},
													{
														"id": "52907745-7672-470e-a803-a2f8feb52944",
														"name": "Entertaint",
														"amount": 4620000
													},
													{
														"id": "52907745-7672-470e-a803-a2f8feb52944",
														"name": "Health",
														"amount": 124000
													}
												]
											},
											"budgets": [
												{
													"id": "52907745-7672-470e-a803-a2f8feb52944",
													"name": "Entertaint",
													"spent": 600000,
													"amount": 1200000
												},
												{
													"id": "52907745-7672-470e-a803-a2f8feb52944",
													"name": "Food",
													"spent": 600000,
													"amount": 1200000
												},
												{
													"id": "52907745-7672-470e-a803-a2f8feb52944",
													"name": "Health",
													"spent": 600000,
													"amount": 1200000
												},
												{
													"id": "52907745-7672-470e-a803-a2f8feb52944",
													"name": "Invest",
													"spent": 600000,
													"amount": 1200000
												}
											]
										}
									}
								}
							}
						}
					}
				},
				"description": "Retrieves the summary data of user finance status",
				"tags": [
					"Dashboard Controllers"
				],
				"security": [],
				"parameters": [
					{
						"description": "Specific time frame for calculating total expenses and incomes",
						"in": "query",
						"name": "timeFrame",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TimeFrame"
						}
					},
					{
						"description": "Time range from",
						"in": "query",
						"name": "from",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Time range to",
						"in": "query",
						"name": "to",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}